+++ build.sh
@@ -13,3 +13,4 @@ cp launcher.dist/* build/ -r
 cp languages/ build/ -r
 cp bannergen/fonts build/ -r
 mkdir build/config
+mv config_sample.json build/config/