+++ src/plugins/announcement.py
@@ -49,6 +49,10 @@ async def getAnnouncement(request: Request, response: Response, authorization: s
     order = order.upper()
 
     if aid != -1:
+        if aid < 0:
+            response.status_code = 404
+            return {"error": True, "descriptor": ml.tr(request, "announcement_not_found")}
+            
         cur.execute(f"SELECT title, content, atype, timestamp, userid, aid, pvt FROM announcement WHERE aid = {aid} {limit}")
         t = cur.fetchall()
         if len(t) == 0:
@@ -184,6 +188,9 @@ async def patchAnnouncement(request: Request, response: Response, authorization:
     if not channelid.isdigit():
         channleid = 0
 
+    if aid < 0:
+        response.status_code = 404
+        return {"error": True, "descriptor": ml.tr(request, "announcement_not_found")}
     cur.execute(f"SELECT userid FROM announcement WHERE aid = {aid}")
     t = cur.fetchall()
     if len(t) == 0:
@@ -232,6 +239,9 @@ async def deleteAnnouncement(aid: int, request: Request, response: Response, aut
         if int(i) in config.perms.admin or int(i) in config.perms.announcement:
             isAdmin = True
 
+    if aid < 0:
+        response.status_code = 404
+        return {"error": True, "descriptor": ml.tr(request, "announcement_not_found")}
     cur.execute(f"SELECT userid FROM announcement WHERE aid = {aid}")
     t = cur.fetchall()
     if len(t) == 0: